Change AWS producer package to com.redhat.swatch.aws
In order to avoid "split package usage"
