+++ swatch-producer-aws/src/main/java/com/redhat/swatch/aws/processors/BillableUsageProducer.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Red Hat trademarks are not licensed under GPLv3. No permission is
+ * granted to use or replicate Red Hat trademarks that are incorporated
+ * in this software or its documentation.
+ */
+package com.redhat.swatch.aws.processors;
+
+import com.redhat.swatch.aws.openapi.model.BillableUsage;
+import jakarta.enterprise.context.ApplicationScoped;
+import lombok.extern.slf4j.Slf4j;
+import org.eclipse.microprofile.reactive.messaging.Channel;
+import org.eclipse.microprofile.reactive.messaging.Emitter;
+
+@Slf4j
+@ApplicationScoped
+public class BillableUsageProducer {
+
+  private final Emitter<BillableUsage> emitter;
+
+  public BillableUsageProducer(@Channel("tally-out") Emitter<BillableUsage> emitter) {
+    this.emitter = emitter;
+  }
+
+  public void queueBillableUsage(BillableUsage billableUsage) {
+    emitter.send(billableUsage);
+
+    log.info("Queued up a BillableUsage object to the billable-usage topic");
+  }
+}
