+++ swatch-producer-aws/src/main/java/com/redhat/swatch/aws/resource/KafkaResource.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * Red Hat trademarks are not licensed under GPLv3. No permission is
+ * granted to use or replicate Red Hat trademarks that are incorporated
+ * in this software or its documentation.
+ */
+package com.redhat.swatch.aws.resource;
+
+import com.redhat.swatch.aws.openapi.model.KafkaSeekPosition;
+import com.redhat.swatch.aws.openapi.resource.KafkaApi;
+import com.redhat.swatch.kafka.KafkaSeekHelper;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.ProcessingException;
+import java.time.OffsetDateTime;
+
+public class KafkaResource implements KafkaApi {
+  private final KafkaSeekHelper kafkaSeekHelper;
+
+  @Inject
+  KafkaResource(KafkaSeekHelper kafkaSeekHelper) {
+    this.kafkaSeekHelper = kafkaSeekHelper;
+  }
+
+  @Override
+  public void kakfaSeekPosition(KafkaSeekPosition position) throws ProcessingException {
+    var seekHelperPosition =
+        switch (position) {
+          case BEGINNING -> KafkaSeekHelper.KafkaSeekPosition.BEGINNING;
+          case END -> KafkaSeekHelper.KafkaSeekPosition.END;
+        };
+    kafkaSeekHelper.seekToPosition(seekHelperPosition);
+  }
+
+  @Override
+  public void kakfaSeekTimestamp(String timestamp) throws ProcessingException {
+    kafkaSeekHelper.seekToTimestamp(OffsetDateTime.parse(timestamp));
+  }
+}
